+++ CHANGELOG.rst
@@ -32,6 +32,7 @@ Added
     - ``ProjectModifyPluginMixin.perform_project_delete()`` method (#1090)
     - ``ProjectDestroyAPIView`` REST API view (#1090)
     - ``ProjectSerializer`` ``children`` field (#1552)
+    - ``SODARUserSerializer`` ``auth_type`` field (#1501)
 
 Changed
 -------